[SERVICE-392] Moved provider UI into 'view' folder, for more consistent MVC structure.
