--- src/provider/Selector.tsx
@@ -1,8 +1,0 @@
-import * as React from 'react';
-import {render} from 'react-dom';
-
-import { AppList } from './components/AppList';
-
-import '../../res/demo/css/w3.css';
-
-render(<AppList />, document.getElementById('react-app'));
