+++ src/provider/view/Selector.tsx
@@ -1,0 +1,8 @@
+import * as React from 'react';
+import {render} from 'react-dom';
+
+import { AppList } from './components/AppList';
+
+import '../../../res/provider/ui/css/w3.css';
+
+render(<AppList />, document.getElementById('react-app'));
