+++ src/provider/view/components/AppCard.css
@@ -1,0 +1,19 @@
+.w3-button.app-card {
+    padding: 2px;
+    width: 100%;
+    text-align: left;
+    margin-bottom: 8px;
+    border: solid lightgray 2px;
+}
+
+.w3-button.app-card.selected {
+    border-color: limegreen;
+}
+
+.app-card img {
+    float: left;
+    padding: 3px;
+    margin: 0 12px;
+    width: 56px;
+    height: 56px;
+}
